Restructured Components into Views, General, and other functionality
