+++ PokeShelter/client/src/components/views/DetailPage.jsx
