+++ PokeShelter/client/src/components/views/Main.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import ResultCard from "./resultcards/resultcard";
-import Nav from "./Nav";
-import Search from "./Search";
+import ResultCard from "../resultcards/resultcard";
+import Nav from "../navsearch/Nav";
+import Search from "../navsearch/Search";
 
 const Main = (props) => {
